make pseudo-wheel part of winpython a true wheel 'wppm'
looks the needed direction towards pure pylock.toml
