+++ make.py
@@ -12,7 +12,7 @@ import shutil
 import subprocess
 import sys
 from pathlib import Path
-from winpython import wppm, utils, diff
+from wppm import wppm, utils, diff
 
 # Define constant paths for clarity
 CHANGELOGS_DIRECTORY = Path(__file__).parent / "changelogs"
@@ -222,7 +222,7 @@ class WinPythonDistributionBuilder:
             utils.python_execmodule("ensurepip", self.distribution.target)
             self.distribution.patch_standard_packages("pip")
 
-            essential_packages = ["pip", "setuptools", "wheel", "winpython"]
+            essential_packages = ["pip", "setuptools", "wheel", "wppm"]
             for package_name in essential_packages:
                 actions = ["install", "--upgrade", "--pre", package_name] + self.install_options
                 self._print_action(f"Piping: {' '.join(actions)}")
@@ -247,8 +247,8 @@ class WinPythonDistributionBuilder:
         diff.write_changelog(self.winpyver2, None, CHANGELOGS_DIRECTORY, self.flavor, self.distribution.architecture, basedir=self.winpython_directory.parent)
 
 def rebuild_winpython_package(source_directory: Path, target_directory: Path, architecture: int = 64, verbose: bool = False):
-    """Rebuilds the winpython package from source using flit."""
-    for file in target_directory.glob("winpython-*"):
+    """Rebuilds the winpython or wppm package from source using flit."""
+    for file in target_directory.glob("w*p*-*.*"):
         if file.suffix in (".exe", ".whl", ".gz"):
             file.unlink()
     utils.buildflit_wininst(source_directory, copy_to=target_directory, verbose=True)
@@ -295,7 +295,7 @@ def make_all(build_number: int, release_level: str, pyver: str, architecture: in
         os.makedirs(build_directory, exist_ok=True)
         # use source_dirs as the directory to re-build Winpython wheel
         winpython_source_dir = Path(__file__).resolve().parent
-        rebuild_winpython_package(winpython_source_dir, Path(source_dirs), architecture, verbose)
+        # 2025-06-28 no more: rebuild_winpython_package(winpython_source_dir, Path(source_dirs), architecture, verbose)
 
     builder = WinPythonDistributionBuilder(
         build_number, release_level, build_directory, wheels_directory=source_dirs,
