+++ wppm/hash.py
